Implement AxROM support; fix address wrapping

+++ dotNES/Mappers/UxROM.cs
@@ -13,7 +13,7 @@
         {
             cpu.MapReadHandler(0x6000, 0x7FFF, addr => _prgRAM[addr - 0x6000]);
             cpu.MapReadHandler(0x8000, 0xBFFF, addr => _prgROM[_bankOffset + (addr - 0x8000)]);
-            cpu.MapReadHandler(0xC000, 0xFFFF, addr => _prgROM[(uint)_prgROM.Length - 0x4000 + (addr - 0xC000)]);
+            cpu.MapReadHandler(0xC000, 0xFFFF, addr => _prgROM[_prgROM.Length - 0x4000 + (addr - 0xC000)]);
 
             cpu.MapWriteHandler(0x6000, 0x7FFF, (addr, val) => _prgRAM[addr - 0x6000] = val);
             cpu.MapWriteHandler(0x8000, 0xFFFF, (addr, val) => _bankOffset = (val & 0xF) * 0x4000);
